executor: split test for aggregate (#44987)
ref pingcap/tidb#44940
