+++ executor/internal/BUILD.bazel
@@ -1,0 +1,9 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "internal",
+    srcs = ["testkit.go"],
+    importpath = "github.com/pingcap/tidb/executor/internal",
+    visibility = ["//executor:__subpackages__"],
+    deps = ["//testkit"],
+)
